refactor: enhance token and auth management with structural updates
Updated `UserAccountService` with email-based account fetching and password verification. Introduced `LoginResult` object for encapsulating responses and cookies. Replaced `Tokens` with `JwtTokens` for better clarity. Standardized cookie building logic in `TokenService` and updated JWT expiration handling.

+++ src/main/java/pl/medflow/medflowbackend/domain/token/JwtProperties.java
@@ -20,9 +20,9 @@ public class JwtProperties {
         private int expirationSeconds;
         private String issuer;
 
-        public long accessExpirationSeconds() {
+        public int accessExpirationSeconds() {
             if (expirationSeconds > 0) return expirationSeconds;
-            return 900L;
+            return 900;
         }
     }
 
@@ -32,9 +32,9 @@ public class JwtProperties {
         private int expirationSeconds;
         private String issuer;
 
-        public long refreshExpirationSeconds() {
+        public int refreshExpirationSeconds() {
             if (expirationSeconds > 0) return expirationSeconds;
-            return 7L * 24 * 60 * 60;
+            return 7 * 24 * 60 * 60;
         }
     }
